feat: implements attribute forms

+++ services/workflow_manager/src/main/java/com/workflowmanager/app/domains/WorkflowAttribute.java
@@ -9,6 +9,7 @@ import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.Instant;
 import java.util.Date;
 import org.springframework.web.server.ResponseStatusException;
@@ -43,7 +44,7 @@ public class WorkflowAttribute {
   }
 
   // values
-  private Long integer;
+  private BigInteger integer;
   private Double floating;
   private String enumeration;
   private BigDecimal decimal;
@@ -71,7 +72,7 @@ public class WorkflowAttribute {
     this.setText(attributeDTO.text);
   }
 
-  public Long getInteger() {
+  public BigInteger getInteger() {
     return this.integer;
   }
 
@@ -103,7 +104,7 @@ public class WorkflowAttribute {
     return this.text;
   }
 
-  public void setInteger(Long integer) throws ResponseStatusException {
+  public void setInteger(BigInteger integer) throws ResponseStatusException {
     if (integer == null) {
       this.integer = null;
       return;
